Modelo 2.0.4 (#1078)
* Update devapp1.yml

New release

* Update playwright-dev.yml

New release

* Update package.json

* dev chgs to reflect new structure

* Bump minimatch from 3.0.4 to 3.1.2 in /studio (#1079)

Bumps [minimatch](https://github.com/isaacs/minimatch) from 3.0.4 to 3.1.2.
- [Release notes](https://github.com/isaacs/minimatch/releases)
- [Changelog](https://github.com/isaacs/minimatch/blob/main/changelog.md)
- [Commits](https://github.com/isaacs/minimatch/compare/v3.0.4...v3.1.2)

---
updated-dependencies:
- dependency-name: minimatch
  dependency-type: indirect
...


* Undo Integrate PlayWright (#1080)

* Bump minimatch from 3.0.4 to 3.1.2 in /studio (#1079)

Bumps [minimatch](https://github.com/isaacs/minimatch) from 3.0.4 to 3.1.2.
- [Release notes](https://github.com/isaacs/minimatch/releases)
- [Changelog](https://github.com/isaacs/minimatch/blob/main/changelog.md)
- [Commits](https://github.com/isaacs/minimatch/compare/v3.0.4...v3.1.2)

---
updated-dependencies:
- dependency-name: minimatch
  dependency-type: indirect
...


* remove the playwright badge

* un integrate playwright

---------


* Issue 1089 Unique email address (#1105)

* make email column unique

* ignore local configurations & secrets

* open stages in new tab (#1106)

* Issue 1097 Open stages in new tab (cont.) (#1108)

* open stages in new tab

* open as new tab from stage management

* Issue 1089 Update existed user email (#1109)

* test update existed email mutation

* raise error when email address is belongs to another user

* fix unsupported operand types for |

* lock sqlalchemy version

* Issue 1093 Stage table actions (#1110)

* quick login for developers on studio

* fix reactivity not working

* update stage inside studio

* update status & visibility mutation

* table visibility & status toggle

* no split layout (#1111)

* scaffold script enhancement (#1112)


* consistent open stage behavior (#1113)

---------

+++ studio/src/utils/common.ts
@@ -10,8 +10,16 @@ export function getSharedConfig(): SharedConfigs {
     const sharedConfig = JSON.parse(localStorage.getItem("configs") ?? "");
     return sharedConfig;
   } catch (error) {
-    localStorage.clear(); // Remove shared auth so that it will ask you to visit Dashboard for login
-    return {} as SharedConfigs;
+    if (import.meta.env.PROD) {
+      localStorage.clear(); // Remove shared auth so that it will ask you to visit Dashboard for login
+    }
+    return {
+      GRAPHQL_ENDPOINT: import.meta.env.VITE_GRAPHQL_ENDPOINT,
+      MQTT_CONNECTION: {},
+      STREAMING: {
+        auth: {},
+      },
+    } as SharedConfigs;
   }
 }
